Conflicts:
	configure
	src/Makefile.in

+++ src/plugins/task/affinity/task_affinity.c
@@ -77,15 +77,12 @@
  * of how this plugin satisfies that application.  SLURM will only load
  * a task plugin if the plugin_type string has a prefix of "task/".
  *
- * plugin_version - an unsigned 32-bit integer giving the version number
- * of the plugin.  If major and minor revisions are desired, the major
- * version number may be multiplied by a suitable magnitude constant such
- * as 100 or 1000.  Various SLURM versions will likely require a certain
- * minimum version for their plugins as this API matures.
+ * plugin_version - an unsigned 32-bit integer containing the Slurm version
+ * (major.minor.micro combined into a single number).
  */
 const char plugin_name[]        = "task affinity plugin";
 const char plugin_type[]        = "task/affinity";
-const uint32_t plugin_version   = 100;
+const uint32_t plugin_version   = SLURM_VERSION_NUMBER;
 
 /*
  * init() is called when the plugin is loaded, before any other functions
@@ -335,13 +332,13 @@ extern int task_p_pre_setuid (stepd_step_rec_t *job)
 			     CPUSET_DIR,
 			     (conf->node_name != NULL)?conf->node_name:"",
 			     job->jobid) > PATH_MAX) {
-			error("cpuset path too long");
+			error("%s: cpuset path too long", __func__);
 			rc = SLURM_ERROR;
 		}
 #else
 		if (snprintf(path, PATH_MAX, "%s/slurm%u",
 			     CPUSET_DIR, job->jobid) > PATH_MAX) {
-			error("cpuset path too long");
+			error("%s: cpuset path too long", __func__);
 			rc = SLURM_ERROR;
 		}
 #endif
@@ -517,7 +514,9 @@ extern int task_p_post_term (stepd_step_rec_t *job, stepd_step_task_info_t *task
 		error("%s: cpuset path too long", __func__);
 		return SLURM_ERROR;
 	}
-	if (rmdir(path) != 0) {
+	/* Only error out if it failed to remove the cpuset dir. The cpuset
+	 * dir may have already been removed by the release_agent. */
+	if (rmdir(path) != 0 && errno != ENOENT) {
 		error("%s: rmdir(%s) failed %m", __func__, path);
 		return SLURM_ERROR;
 	}
